Updated note detection.
There seems to be an issue with the software FFT such that the low frequency bins of the FFT are much higher than the other FFT bins. That makes things unpredictable.

I don't know if this is purely an issue with software FFT, or if it is an issue with the guitars resonance. I will try with pure tones.

I was unable to implement the ANN like I wanted due to being unable to get the note detection algorithm working with the software FFT.